normchecks

+++ ft_calloc.c
@@ -10,19 +10,20 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-//Allocates to heap memory and returns a pointer. fills pointer with zeros.
+//Allocates to heap memory and returns a pointer. 
+// fills pointer with zeros.
 
 #include "libft.h"
 
-void    *ft_calloc(size_t count, size_t size)
+void	*ft_calloc(size_t count, size_t size)
 {
-    void    *ptr;
-    int len;
+	void	*ptr;
+	size_t	len;
 
-    len = size * count;
-    ptr = malloc(len);
-    if(!ptr)
-        return (NULL);
-    ft_bzero(ptr, len);
-    return (ptr);
+	len = size * count;
+	ptr = malloc(len);
+	if (!ptr)
+		return (NULL);
+	ft_bzero(ptr, len);
+	return (ptr);
 }
